% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{altitude_to_numeric}
\alias{altitude_to_numeric}
\title{Convert altitude column to numeric}
\usage{
altitude_to_numeric(df)
}
\arguments{
\item{df}{A data.frame internal to the \code{read_airport()} function.}
}
\value{
A \verb{"data.table" "data.frame"} object
}
\description{
Convert altitude column to numeric
}
\keyword{internal}
