% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_date}
\alias{check_date}
\title{Check whether date input is acceptable}
\usage{
check_date(date, all_dates)
}
\arguments{
\item{date}{Numeric. Either a 6-digit date in the format \code{yyyymm} or a 4-digit
date input \code{yyyy} .}

\item{all_dates}{Numeric vector created with the get_all_dates_available() function.}
}
\value{
Check messages.
}
\description{
Check whether date input is acceptable
}
\examples{
\dontrun{ if (interactive()) {

# get all dates available
all_dates <- get_all_dates_available()

# check dates
a <- check_date(200011, all_dates)
}}
}
\keyword{internal}
