% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_input_date_format}
\alias{check_input_date_format}
\title{Check whether the format of the date input is acceptable}
\usage{
check_input_date_format(date = parent.frame()$date)
}
\arguments{
\item{date}{Vector. Either a 6-digit date in the format \code{yyyymm} or a 4-digit
date input \code{yyyy} .}
}
\value{
Check messages.
}
\description{
Check whether the format of the date input is acceptable
}
\examples{
\dontrun{ if (interactive()) {

# get all dates available
all_dates <- get_all_dates_available()

# check dates
a <- check_date(200011, all_dates)
}}
}
\keyword{internal}
