% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{download_file}
\alias{download_file}
\title{Download file from url}
\usage{
download_file(file_url, showProgress = showProgress)
}
\arguments{
\item{file_url}{String. A url passed from get_flights_url.}

\item{showProgress}{Logical, passed from \code{\link{read_flights}}}
}
\value{
Silently saves downloaded file to temp dir.
}
\description{
Download file from url
}
\examples{
\dontrun{ if (interactive()) {
# Generate url
file_url <- get_flights_url(type='basica', year=2000, month=11)

# download data
download_file(file_url=file_url, showProgress=TRUE)
}}
}
\keyword{internal}
