% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{download_flightsbr_file}
\alias{download_flightsbr_file}
\title{Download file from url}
\usage{
download_flightsbr_file(
  file_url = parent.frame()$file_url,
  showProgress = parent.frame()$showProgress,
  dest_file = temp_local_file,
  cache = cache
)
}
\arguments{
\item{file_url}{String. A url passed from above.}

\item{showProgress}{Logical, passed from above.}

\item{dest_file}{String, passed from above.}

\item{cache}{Logical, passed from above.}
}
\value{
Silently saves downloaded file to temp dir.
}
\description{
Download file from url
}
\examples{
\dontrun{ if (interactive()) {
# Generate url
file_url <- get_flights_url(type='basica', date=200011)

# download data
download_flightsbr_file(file_url=file_url,
                        showProgress=TRUE,
                        dest_file = tempfile(fileext = ".zip")
                       )
}}
}
\keyword{internal}
