% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_air_fares_dates_available}
\alias{get_air_fares_dates_available}
\title{Retrieve all dates available for air fares data from ANAC website}
\usage{
get_air_fares_dates_available()
}
\value{
Numeric vector.
}
\description{
Retrieve all dates available for air fares data from ANAC website
}
\examples{
\dontrun{ if (interactive()) {
# check dates
a <- get_air_fares_dates_available()
}}
}
\keyword{internal}
