% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_air_fares_url}
\alias{get_air_fares_url}
\title{Put together the url of air fare data files}
\usage{
get_air_fares_url(type, year, month)
}
\arguments{
\item{year}{Numeric. Year of the data in \code{yyyy} format.}

\item{month}{Numeric. Month of the data in \code{mm} format.}
}
\value{
A url string.
}
\description{
Put together the url of air fare data files
}
\examples{
\dontrun{ if (interactive()) {
# Generate url
a <- get_air_fares_url(year=2002, month=11)
}}
}
\keyword{internal}
