% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_aircraft.R
\name{get_aircraft_dates_available}
\alias{get_aircraft_dates_available}
\title{Retrieve all dates available for aircraft data from ANAC website}
\usage{
get_aircraft_dates_available()
}
\value{
Numeric vector.
}
\description{
Retrieve all dates available for aircraft data from ANAC website
}
\examples{
\dontrun{ if (interactive()) {
# check dates
a <- get_aircraft_dates_available()
}}
}
\keyword{internal}
