% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_aircrafts_url}
\alias{get_aircrafts_url}
\title{Put together the url of aircrafts data files}
\usage{
get_aircrafts_url(year, month)
}
\arguments{
\item{year}{Numeric. Year of the data in \code{yyyy} format.}

\item{month}{Numeric. Month of the data in \code{mm} format.}
}
\value{
A url string.
}
\description{
Put together the url of aircrafts data files
}
\examples{
\dontrun{ if (interactive()) {
# Generate url
a <- get_flights_url(year=2000, month=11)
}}
}
\keyword{internal}
