% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_airport_movement.R
\name{get_airport_movement_dates_available}
\alias{get_airport_movement_dates_available}
\title{Retrieve all dates available for airport movements data}
\usage{
get_airport_movement_dates_available()
}
\value{
Numeric vector.
}
\description{
Retrieve from ANAC website all dates available for airport movements data.
}
\examples{
\dontrun{ if (interactive()) {
# check dates
a <- get_airport_movement_dates_available()
}}
}
\keyword{internal}
