% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_all_dates_available}
\alias{get_all_dates_available}
\title{Retrieve from ANAC website all dates available for flights data}
\usage{
get_all_dates_available()
}
\value{
Numeric vector.
}
\description{
Retrieve from ANAC website all dates available for flights data
}
\examples{
\dontrun{ if (interactive()) {
# check dates
a <- get_all_dates_available()
}}
}
\keyword{internal}
