% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_flights.R
\name{get_flight_dates_available}
\alias{get_flight_dates_available}
\title{Retrieve all dates available for flights data from ANAC website}
\usage{
get_flight_dates_available(type = NULL)
}
\arguments{
\item{type}{String. Whether the data set should be of the type \code{basica}
(flight stage, the default) or \code{combinada} (On flight origin and
destination - OFOD).}
}
\value{
Numeric vector.
}
\description{
Retrieve all dates available for flights data from ANAC website
}
\examples{
\dontrun{ if (interactive()) {
# check dates
a <- get_flight_dates_available()
}}
}
\keyword{internal}
