% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_flights_url}
\alias{get_flights_url}
\title{Put together the url of flight data files}
\usage{
get_flights_url(type, date)
}
\arguments{
\item{type}{String. Whether the data set should be of the type \code{basica}
(flight stage, the default) or \code{combinada} (On flight origin and
destination - OFOD).}

\item{date}{Numeric. Date of the data in the format \code{yyyymm}. Defaults to
\code{202001}. To download the data for all months in a year, the user
can pass a 4-digit year input \code{yyyy}. The parameter also accepts
a vector of dates such as \code{c(202001, 202006, 202012)}.}
}
\value{
A url string.
}
\description{
Put together the url of flight data files
}
\examples{
\dontrun{ if (interactive()) {
# Generate urls
a <- get_flights_url(type='basica', year=2000, month=11)
}}
}
\keyword{internal}
