% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latest_flights_date.R
\name{latest_flights_date}
\alias{latest_flights_date}
\title{Check the date of the latest flight data available}
\usage{
latest_flights_date()
}
\value{
A numeric date in the format \code{yyyymm}.
}
\description{
Check the date of the latest flight data available
}
\examples{
\dontrun{ if (interactive()) {

latest_date <- latest_flights_date()

}}
}
\seealso{
Other support function:
\code{\link[=latest_airfares_date]{latest_airfares_date()}}
}
\concept{support function}
