% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{latlon_to_numeric}
\alias{latlon_to_numeric}
\title{Convert latitude and longitude columns to numeric}
\usage{
latlon_to_numeric(df)
}
\arguments{
\item{df}{A data.frame internal to the \code{read_airport()} function.}
}
\value{
A \verb{"data.table" "data.frame"} object
}
\description{
Convert latitude and longitude columns to numeric
}
\keyword{internal}
