% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_air_fares.R
\name{read_air_fares}
\alias{read_air_fares}
\title{Download data on air fares of domestic flights in Brazil}
\usage{
read_air_fares(date = 202001, showProgress = TRUE, select = NULL)
}
\arguments{
\item{date}{Numeric. Date of the data in the format \code{yyyymm}. Defaults to
\code{202001}. To download the data for all months in a year, the user
can pass a 4-digit year input \code{yyyy}.}

\item{showProgress}{Logical. Defaults to \code{TRUE} display progress.}

\item{select}{A vector of column names or numbers to keep, drop the rest. The
order that the columns are specified determines the order of the
columns in the result.}
}
\value{
A \verb{"data.table" "data.frame"} object. All columns are returned with
\code{class} of type \code{"character"}.
}
\description{
Download data on air fares of domestic flights in Brazil. The data is
collected by Brazil’s Civil Aviation Agency (ANAC). A description of all
variables included in the data is available at \url{https://www.anac.gov.br/acesso-a-informacao/dados-abertos/areas-de-atuacao/voos-e-operacoes-aereas/tarifas-aereas-domesticas/46-tarifas-aereas-domesticas}.
}
\examples{
\dontrun{ if (interactive()) {
# Read air fare data
af_201506 <- read_air_fares(date = 201506)

af_2015 <- read_air_fares(date = 2015)
}}
}
\concept{download air fares data}
