% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_airport_operations.R
\name{read_airport_operations}
\alias{read_airport_operations}
\title{Download airport operation data from Brazil}
\usage{
read_airport_operations(date = 202001, showProgress = TRUE)
}
\arguments{
\item{date}{Numeric. Date of the data in the format \code{yyyymm}. Defaults to
\code{202001}. To download the data for all months in a year, the user
can pass a 4-digit year input \code{yyyy}.}

\item{showProgress}{Logical. Defaults to \code{TRUE} display progress.}
}
\value{
A \verb{"data.table" "data.frame"} object.
}
\description{
Download airport operation data from Brazil’s Civil Aviation Agency (ANAC)...
}
\examples{
\dontrun{ if (interactive()) {
# Read flights data
a201506 <- read_airport_operations(date = 202006)

a2015 <- read_airport_operations(date = 2020)
}}
}
\concept{download airport operation data}
