% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_airports.R
\name{read_airports}
\alias{read_airports}
\title{Download airports data from Brazil}
\usage{
read_airports(type = "public", showProgress = TRUE)
}
\arguments{
\item{type}{String. Whether function should download data  on \code{public} (Default)
or \code{private} airports.}

\item{showProgress}{Logical. Defaults to \code{TRUE} display progress.}
}
\value{
A \verb{"data.table" "data.frame"} object.
}
\description{
Download data of all airports and aerodromes registered in Brazil’s Civil
Aviation Agency (ANAC). Data source: \url{https://www.gov.br/anac/pt-br/assuntos/regulados/aerodromos/lista-de-aerodromos-civis-cadastrados}.
The data dictionary for public airports can be found at \url{https://www.anac.gov.br/acesso-a-informacao/dados-abertos/areas-de-atuacao/aerodromos/lista-de-aerodromos-publicos-v2/70-lista-de-aerodromos-publicos-v2}.
The data dictionary for private airports can be found at \url{https://www.anac.gov.br/acesso-a-informacao/dados-abertos/areas-de-atuacao/aerodromos/lista-de-aerodromos-privados-v2}.
}
\examples{
\dontrun{ if (interactive()) {
# Read airports data
public_airports <- read_airports(type = 'public')

private_airports <- read_airports(type = 'private')
}}
}
\concept{download airport data}
