% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_flights.R
\name{read_flights}
\alias{read_flights}
\title{Download flight data from Brazil}
\usage{
read_flights(
  date = NULL,
  type = c("basica", "combinada"),
  showProgress = TRUE,
  select = NULL,
  cache = TRUE
)
}
\arguments{
\item{date}{Numeric. Date of the data in the format \code{yyyymm}. Defaults to
the latest available month. To download the data for all months
in a year, the user can pass a 4-digit year input \code{yyyy}. The
parameter also accepts a vector of dates such as
\code{c(202001, 202006, 202012)}.}

\item{type}{String. Whether the data set should be of the type \code{basica}
(flight stage, the default) or \code{combinada} (On flight origin and
destination - OFOD).}

\item{showProgress}{Logical. Defaults to \code{TRUE} display progress.}

\item{select}{A vector of column names or positions to keep. The rest of the
columns are not read. The order that the columns passed determines the
order of the columns in the result.}

\item{cache}{Logical. Whether the function should read cached data downloaded
previously. Defaults to \code{TRUE}. If \code{FALSE}, the function will
always download the data and overwrite cached data.}
}
\value{
A \verb{"data.table" "data.frame"} object. All columns are returned with
\code{class} of type \code{"character"}.
}
\description{
Download flight data from Brazil’s Civil Aviation Agency (ANAC). The data
includes detailed information on every international flight to and from Brazil,
as well as domestic flights within the country. The data include flight-level
information of airports of origin and destination, flight duration, aircraft
type, payload, and the number of passengers, and several other variables. A
description of all variables included in the data is available at \url{https://www.gov.br/anac/pt-br/assuntos/regulados/empresas-aereas/Instrucoes-para-a-elaboracao-e-apresentacao-das-demonstracoes-contabeis/descricao-de-variaveis}.
}
\examples{
\dontrun{ if (interactive()) {
# Read flights data
f201506 <- read_flights(date = 201506)

f2015 <- read_flights(date = 2015)
}}
}
\seealso{
Other download flight data:
\code{\link[=read_aircraft]{read_aircraft()}},
\code{\link[=read_aircrafts]{read_aircrafts()}}
}
\concept{download flight data}
