% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{split_date}
\alias{split_date}
\title{Split a date from yyyymmm to year yyyy and month mm}
\usage{
split_date(date)
}
\arguments{
\item{date}{Numeric. Date of the data in the format \code{yyyymm}.}
}
\value{
An two string objects, \code{year} and \code{month}.
}
\description{
Split a date from yyyymmm to year yyyy and month mm
}
\keyword{internal}
