% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_date}
\alias{check_date}
\title{Check whether date input is acceptable}
\usage{
check_date(date)
}
\arguments{
\item{date}{Numeric. Either a 6-digit date in the format \code{yyyymm} or a 4-digit
date input \code{yyyy} .}
}
\value{
Check messages.
}
\description{
Check whether date input is acceptable
}
\keyword{internal}
