% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_flights.R
\name{download_flights_data}
\alias{download_flights_data}
\title{Download and read ANAC flight data}
\usage{
download_flights_data(
  file_url = parent.frame()$file_url,
  showProgress = parent.frame()$showProgress,
  select = parent.frame()$select,
  cache = parent.frame()$cache
)
}
\arguments{
\item{file_url}{String. An ANAC url passed from above.}

\item{showProgress}{Logical, passed from above.}

\item{select}{A vector of column names or numbers to keep, passed from above.}

\item{cache}{Logical, passed from above.}
}
\value{
A \verb{"data.table" "data.frame"} object
}
\description{
Download and read ANAC flight data
}
\examples{
\dontrun{ if (interactive()) {
# Generate url
file_url <- get_flights_files_available()$url[1]

# download data
a <- download_flights_data(
  file_url=file_url,
  showProgress=TRUE,
  select=NULL,
  cache=TRUE
  )
}}
}
\keyword{internal}
