% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{download_flights_data}
\alias{download_flights_data}
\title{Download and read ANAC flight data}
\usage{
download_flights_data(file_url, showProgress = showProgress, select = select)
}
\arguments{
\item{file_url}{String. A url passed from get_url.}

\item{showProgress}{Logical, passed from \code{\link{read_flights}}}

\item{select}{A vector of column names or numbers to keep, passed from \code{\link{read_flights}}}
}
\value{
A \verb{"data.table" "data.frame"} object
}
\description{
Download and read ANAC flight data
}
\keyword{internal}
