% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flightsbr.R
\docType{package}
\name{flightsbr}
\alias{flightsbr}
\alias{flightsbr-package}
\title{flightsbr: Download Flight and Airport Data from Brazil}
\description{
Download flight and airport data from Brazil’s Civil Aviation Agency (ANAC)
\url{https://www.gov.br/anac/pt-br}. The data includes detailed information on all
aircrafts, aerodromes, airports, and airport movements registered in ANAC,
and on every international flight to and from Brazil, as well as domestic
flights within the country.
}
\section{Usage}{

Please check the vignettes and data documentation on the
\href{https://ipea.github.io/flightsbr/}{website}.
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ipea/flightsbr}
  \item Report bugs at \url{https://github.com/ipea/flightsbr/issues}
}

}
\author{
\strong{Maintainer}: Rafael H. M. Pereira \email{rafa.pereira.br@gmail.com} (\href{https://orcid.org/0000-0003-2125-7465}{ORCID})

Authors:
\itemize{
  \item Rafael H. M. Pereira \email{rafa.pereira.br@gmail.com} (\href{https://orcid.org/0000-0003-2125-7465}{ORCID})
}

Other contributors:
\itemize{
  \item Arthur Bazolli \email{baz.arthur@gmail.com} [contributor]
  \item Ipea - Institute for Applied Economic Research [copyright holder, funder]
}

}
