% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{generate_all_months}
\alias{generate_all_months}
\title{Generate all months with \code{yyyymm} format in a year}
\usage{
generate_all_months(date)
}
\arguments{
\item{date}{Numeric. 4-digit date in the format \code{yyyy}.}
}
\value{
Vector or strings.
}
\description{
Generate all months with \code{yyyymm} format in a year
}
\keyword{internal}
