% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_aircrafts.R
\name{get_aircrafts_url}
\alias{get_aircrafts_url}
\title{Put together the url of aircrafts data files}
\usage{
get_aircrafts_url(date = parent.frame()$date)
}
\arguments{
\item{date}{Numeric. Either a 6-digit date in the format \code{yyyymm} or a 4-digit
date input \code{yyyy}. Defaults to \code{NULL}, in which case the function
retrieves information for all years available.}
}
\value{
A url string.
}
\description{
Put together the url of aircrafts data files
}
\examples{
\dontrun{ if (interactive()) {
# Generate url
a <- get_flights_url(year=2000, month=11)
}}
}
\keyword{internal}
