% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_airfares.R
\name{get_airfares_dates_available}
\alias{get_airfares_dates_available}
\title{Retrieve all dates available for airfares data from ANAC website}
\usage{
get_airfares_dates_available(dom)
}
\arguments{
\item{dom}{Logical. Defaults to \code{TRUE} download airfares of domestic
flights. If \code{FALSE}, the function downloads airfares of
international flights.}
}
\value{
Numeric vector.
}
\description{
Retrieve all dates available for airfares data from ANAC website
}
\examples{
\dontrun{ if (interactive()) {
# check dates
a <- get_airfares_dates_available(domestic = TRUE)
}}
}
\keyword{internal}
