% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_airfares_url}
\alias{get_airfares_url}
\title{Put together the url of airfare data files}
\usage{
get_airfares_url(dom, year, month)
}
\arguments{
\item{dom}{Logical. Defaults to \code{TRUE} download airfares of domestic
flights. If \code{FALSE}, the function downloads airfares of
international flights.}

\item{year}{Numeric. Year of the data in \code{yyyy} format.}

\item{month}{Numeric. Month of the data in \code{mm} format.}
}
\value{
A url string.
}
\description{
Put together the url of airfare data files
}
\examples{
\dontrun{ if (interactive()) {
# Generate url
a <- get_airfares_url(year=2002, month=11)
}}
}
\keyword{internal}
