% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_airport_movement_dates_available}
\alias{get_airport_movement_dates_available}
\title{Retrieve all dates available for airport movements data}
\usage{
get_airport_movement_dates_available(date = NULL)
}
\arguments{
\item{date}{Numeric. Either a 6-digit date in the format \code{yyyymm} or a 4-digit
date input \code{yyyy}. Defaults to \code{NULL}, in which case the function
retrieves information for all years available.}
}
\value{
Numeric vector.
}
\description{
Retrieve from ANAC website all dates available for airport movements data.
}
\examples{
\dontrun{ if (interactive()) {
# check dates
a <- get_airport_movement_dates_available()
}}
}
\keyword{internal}
