% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_airport_movement.R
\name{get_airport_movements_url}
\alias{get_airport_movements_url}
\title{Put together the url of airport movement data files}
\usage{
get_airport_movements_url(date = parent.frame()$date)
}
\arguments{
\item{date}{Numeric. Date of the data in the format \code{yyyymm}. Defaults to
\code{202001}. To download the data for all months in a year, the user
can pass a 4-digit year input \code{yyyy}. The parameter also accepts
a vector of dates such as \code{c(202001, 202006, 202012)}.}
}
\value{
A url string.
}
\description{
Put together the url of airport movement data files
}
\examples{
\dontrun{ if (interactive()) {
# Generate url
a <- get_airport_movements_url(year=2000, month=11)
}}
}
\keyword{internal}
