% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_flights.R
\name{get_flights_files_available}
\alias{get_flights_files_available}
\title{Retrieve flight files available from the ANAC website}
\usage{
get_flights_files_available(cache = TRUE)
}
\arguments{
\item{cache}{Logical. Whether the function should read cached data downloaded
previously. Defaults to \code{TRUE}. If \code{FALSE}, the function will
always download the data and overwrite cached data.}
}
\value{
A data.table with columns \code{date}, \code{type}, and \code{url}.
}
\description{
Retrieve flight files available from the ANAC website
}
\keyword{internal}
