% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latest_flights_date.R
\name{latest_airfares_date}
\alias{latest_airfares_date}
\title{Check the date of the latest airfare data available}
\usage{
latest_airfares_date(dom = TRUE)
}
\arguments{
\item{dom}{Logical. Defaults to \code{TRUE} download airfares of domestic
flights. If \code{FALSE}, the function downloads airfares of
international flights.}
}
\value{
A numeric date in the format \code{yyyymm}.
}
\description{
Check the date of the latest airfare data available
}
\examples{
\dontrun{ if (interactive()) {

latest_date <- latest_airfares_date()

}}
}
\seealso{
Other support function: 
\code{\link{latest_flights_date}()}
}
\concept{support function}
