% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_aircrafts.R
\name{read_aircrafts}
\alias{read_aircrafts}
\title{Download aircraft data from Brazil}
\usage{
read_aircrafts(showProgress = TRUE)
}
\arguments{
\item{showProgress}{Logical. Defaults to \code{TRUE} display progress.}
}
\value{
A \verb{"data.table" "data.frame"} object.
}
\description{
Download data of all aircrafts registered in the Brazilian Aeronautical
Registry (Registro Aeronáutico Brasileiro - RAB), organized by the Brazilian
Civil Aviation Agency (ANAC). A description of all variables included in the
data is available at \url{https://dados.gov.br/dataset/aeronaves-registradas-no-registro-aeronautico-brasileiro-rab}.
}
\examples{
\dontrun{ if (interactive()) {
# Read aircrafts data
aircrafts <- read_aircrafts(showProgress = TRUE)

}}
}
\seealso{
Other download flight data: 
\code{\link{read_flights}()}
}
\concept{download flight data}
